Refactor the `tools pkg repo` commands into a subdirectory
