+++ tools/__init__.py
@@ -8,6 +8,8 @@ ptscripts.register_tools_module("tools.docs")
 ptscripts.register_tools_module("tools.pkg")
 ptscripts.register_tools_module("tools.pkg.repo")
 ptscripts.register_tools_module("tools.pkg.build")
+ptscripts.register_tools_module("tools.pkg.repo.create")
+ptscripts.register_tools_module("tools.pkg.repo.publish")
 ptscripts.register_tools_module("tools.pre_commit")
 ptscripts.register_tools_module("tools.release")
 ptscripts.register_tools_module("tools.vm")
